Enables include-src formation with makefile support

+++ src/sampler.cpp
@@ -2,21 +2,20 @@
 #include <iostream>
 #include <cmath>
 #include <vector>
-#include "system.h"
-#include "sampler.h"
-#include "particle.h"
-#include "Hamiltonians/hamiltonian.h"
-#include "WaveFunctions/wavefunction.h"
+#include "../include/system.h"
+#include "../include/sampler.h"
+#include "../include/particle.h"
+#include "../include/hamiltonian.h"
+#include "../include/wavefunction.h"
 
 using std::cout;
 using std::endl;
 
-
 Sampler::Sampler(
-        unsigned int numberOfParticles,
-        unsigned int numberOfDimensions,
-        double stepLength,
-        unsigned int numberOfMetropolisSteps)
+    size_t numberOfParticles,
+    size_t numberOfDimensions,
+    double stepLength,
+    size_t numberOfMetropolisSteps)
 {
     m_stepNumber = 0;
     m_numberOfMetropolisSteps = numberOfMetropolisSteps;
@@ -28,18 +27,19 @@ Sampler::Sampler(
     m_numberOfAcceptedSteps = 0;
 }
 
-
-void Sampler::sample(bool acceptedStep, System* system) {
+void Sampler::sample(bool acceptedStep, System *system)
+{
     /* Here you should sample all the interesting things you want to measure.
      * Note that there are (way) more than the single one here currently.
      */
     auto localEnergy = system->computeLocalEnergy();
-    m_cumulativeEnergy  += localEnergy;
+    m_cumulativeEnergy += localEnergy;
     m_stepNumber++;
     m_numberOfAcceptedSteps += acceptedStep;
 }
 
-void Sampler::printOutputToTerminal(System& system) {
+void Sampler::printOutputToTerminal(System &system)
+{
     auto pa = system.getWaveFunctionParameters();
     auto p = pa.size();
 
@@ -49,12 +49,13 @@ void Sampler::printOutputToTerminal(System& system) {
     cout << " Number of dimensions : " << m_numberOfDimensions << endl;
     cout << " Number of Metropolis steps run : 10^" << std::log10(m_numberOfMetropolisSteps) << endl;
     cout << " Step length used : " << m_stepLength << endl;
-    cout << " Ratio of accepted steps: " << ((double) m_numberOfAcceptedSteps) / ((double) m_numberOfMetropolisSteps) << endl;
+    cout << " Ratio of accepted steps: " << ((double)m_numberOfAcceptedSteps) / ((double)m_numberOfMetropolisSteps) << endl;
     cout << endl;
     cout << "  -- Wave function parameters -- " << endl;
     cout << " Number of parameters : " << p << endl;
-    for (unsigned int i=0; i < p; i++) {
-        cout << " Parameter " << i+1 << " : " << pa.at(i) << endl;
+    for (size_t i = 0; i < p; i++)
+    {
+        cout << " Parameter " << i + 1 << " : " << pa.at(i) << endl;
     }
     cout << endl;
     cout << "  -- Results -- " << endl;
@@ -62,7 +63,8 @@ void Sampler::printOutputToTerminal(System& system) {
     cout << endl;
 }
 
-void Sampler::computeAverages() {
+void Sampler::computeAverages()
+{
     /* Compute the averages of the sampled quantities.
      */
     m_energy = m_cumulativeEnergy / m_numberOfMetropolisSteps;
